airflow/music_app/backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
openai
pydantic
pandas
db-dtypes